Outline of our frontend
This first pr includes the navigation bar for switch between party finder, guides (or whatever this may be in the future), and friends. This also includes a dropdown menu for profile, settings, and logout. None of these buttons are functional at the moment but will be added in a future pr.

Also add was adding a button for create, a search bar, and a button to filter. At this moment, of the create button has any functionality.|

The create button current allows the user to create a party by applying options that include party type, party size, and an optional description/note.

Once create, the party will be added to the list of parties and will show the type, amount of players as boxes, the description (displays No description if the field was left empty on creation), a number visual for current party with party size, and a button to join the party.

Currently, the boxes change from yellow to black when join is clicked button will display the skins of players instead. The join currently can be clicked as many times as the user wants but will be changed in the future once users are implemented into groups. Finally if the party is full, the join button changes into text saying the party is full.